Moved contents from src/clocks/mod.rs to src/clocks/clocks.rs
